requirements.txt
ScrapyElasticSearch==0.8.3
requests_ntlm==0.2.0